Moved demo code to separate file.
